requirements.txt
openpyxl
pandas
numpy
matplotlib
Pillow
names
svglib
reportlab